GuiPanelTexture: migrate uv-texture to component, overlays in parent, rmv deprecated classes
