+++ src/components/panel/textures/GuiPanelTexture.tsx
@@ -1,22 +1,30 @@
 import clsx from 'clsx';
-import { useContext, useEffect, useMemo, useRef, useState } from 'react';
-import { Box, ButtonBase, Skeleton, Tooltip } from '@mui/material';
+import {
+  MouseEvent,
+  useCallback,
+  useContext,
+  useMemo,
+  useRef,
+  useState
+} from 'react';
+import AccountTreeIcon from '@mui/icons-material/AccountTree';
+import MoreVertIcon from '@mui/icons-material/MoreVert';
+import { Box, ButtonBase, IconButton, Skeleton, Tooltip } from '@mui/material';
 import type { SxProps, Theme } from '@mui/material/styles';
 import type { ContentViewMode, NLUITextureDef } from '@/types';
 import GuiPanelTextureMenu from './GuiPanelTextureMenu';
 import GuiPanelTextureModelNavMenu from './GuiPanelTextureModelNavMenu';
+import GuiPanelTextureUvPreview, {
+  type ClipPathGroup
+} from './GuiPanelTextureUvPreview';
 import type { TextureModelReference } from '@/modules/model-data';
 import { selectModel, selectModels } from '@/selectors';
 import { setObjectViewedIndex } from '@/modules/object-viewer';
 import { useAppDispatch, useAppSelector } from '@/storeTypings';
 import SceneOptionsContext from '@/contexts/SceneOptionsContext';
 import { useTextureReplaceDropzone } from '@/modules/replace-texture';
-import ImageBufferCanvas from '@/components/ImageBufferCanvas';
-import globalBuffers from '@/utils/data/globalBuffers';
 import { createUvClipPaths, UvClipPath } from '@/utils/textures';
 
-const HOVERED_MODEL_UV_PULSE_MS = 2000;
-const HOVERED_MODEL_UV_MAX_ALPHA = 0.275;
 const HOVERED_MODEL_UV_COLORS = [
   [255, 209, 102],
   [6, 214, 160],
@@ -57,35 +65,17 @@ const panelTextureSx: SxProps<Theme> = (theme) => ({
   '&.mode-textures.selectable': {
     cursor: 'pointer'
   },
-  '& .uv-overlay': {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    width: '100%',
-    height: '100%',
-    border: '3px solid transparent',
-    pointerEvents: 'none'
+  '& .texture-overlay-button': {
+    opacity: 0,
+    transition: theme.transitions.create('opacity', {
+      duration: theme.transitions.duration.short
+    })
   },
-  '& .uv-overlay canvas': {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    width: '100%',
-    height: '100%',
-    zIndex: 1,
-    transform: 'rotate(-90deg)',
-    pointerEvents: 'none'
-  },
-  '& .size-notation': {
-    position: 'absolute',
-    right: 'var(--mui-spacing)',
-    bottom: 'var(--mui-spacing)',
+  '& .texture-overlay-button svg': {
     color: 'var(--mui-palette-primary-contrastText)',
-    textShadow: '1px 1px 1px black',
-    filter: 'drop-shadow(3px 3px 1px black)',
-    userSelect: 'none'
+    filter: 'drop-shadow(3px 5px 2px rgb(0 0 0 / 0.8))'
   },
-  '&:hover .MuiIconButton-root': {
+  '&:hover .texture-overlay-button': {
     opacity: 1
   }
 });
@@ -108,11 +98,6 @@ export type GuiPanelTextureProps =
       selectedTextureReferences: undefined;
       contentViewMode: undefined;
     };
-
-interface ClipPathGroup {
-  paths: UvClipPath[];
-  color?: Readonly<[r: number, g: number, b: number]>;
-}
 
 // rotate in 90 deg to match the orientation of stored textures
 const rotateUvClipPathForTextureEdit = (
@@ -168,13 +153,36 @@ export default function GuiPanelTexture(props: GuiPanelTextureProps) {
     contentViewMode = 'textures'
   } = props;
   const dispatch = useAppDispatch();
+  const textureContainerRef = useRef<HTMLDivElement>(null);
   const model = useAppSelector(selectModel);
   const models = useAppSelector(selectModels);
   const [hoveredModelReference, setHoveredModelReference] =
     useState<TextureModelReference>();
+  const [textureMenuAnchorEl, setTextureMenuAnchorEl] =
+    useState<HTMLElement | null>(null);
+  const [modelNavMenuAnchorEl, setModelNavMenuAnchorEl] =
+    useState<HTMLElement | null>(null);
   const { textureViewMode, uvRegionsHighlighted } =
     useContext(SceneOptionsContext);
   const { width = 0, height = 0 } = textureDef || {};
+  const textureMenuOpen = Boolean(textureMenuAnchorEl);
+  const modelNavMenuOpen = Boolean(modelNavMenuAnchorEl);
+
+  const onOpenTextureMenu = useCallback((event: MouseEvent<HTMLElement>) => {
+    setTextureMenuAnchorEl(textureContainerRef.current ?? event.currentTarget);
+  }, []);
+
+  const onCloseTextureMenu = useCallback(() => {
+    setTextureMenuAnchorEl(null);
+  }, []);
+
+  const onOpenModelNavMenu = useCallback((event: MouseEvent<HTMLElement>) => {
+    setModelNavMenuAnchorEl(textureContainerRef.current ?? event.currentTarget);
+  }, []);
+
+  const onCloseModelNavMenu = useCallback(() => {
+    setModelNavMenuAnchorEl(null);
+  }, []);
 
   const hoveredUvClipPathGroups = useMemo<ClipPathGroup[]>(() => {
     if (!textureDef) {
@@ -276,12 +284,6 @@ export default function GuiPanelTexture(props: GuiPanelTextureProps) {
       : textureDef?.bufferKeys?.translucent ||
         textureDef?.bufferKeys?.opaque) || '';
 
-  const rgbaBuffer = useMemo(
-    () =>
-      imageBufferKey ? globalBuffers.get(imageBufferKey) : new Uint8Array(0),
-    [imageBufferKey]
-  );
-
   const showUvHighlight = Boolean(
     hasUvClipPaths &&
     (hoveredModelReference || (uvRegionsHighlighted && selected))
@@ -292,139 +294,6 @@ export default function GuiPanelTexture(props: GuiPanelTextureProps) {
   const uvHighlightBufferKey = showUvHighlight
     ? textureDef?.bufferKeys?.opaque
     : undefined;
-  const shouldUseSeparateUvHighlightBuffer =
-    uvHighlightBufferKey && uvHighlightBufferKey !== imageBufferKey;
-  const uvHighlightRgbaBuffer = useMemo(
-    () =>
-      shouldUseSeparateUvHighlightBuffer
-        ? globalBuffers.get(uvHighlightBufferKey)
-        : undefined,
-    [shouldUseSeparateUvHighlightBuffer, uvHighlightBufferKey]
-  );
-
-  const [textureBitmaps, setTextureBitmaps] = useState<{
-    source: ImageBitmap | null;
-    highlight: ImageBitmap | null;
-  }>({ source: null, highlight: null });
-
-  useEffect(() => {
-    let active = true;
-
-    Promise.all(
-      [rgbaBuffer, uvHighlightRgbaBuffer].map((buffer) => {
-        if (
-          !buffer?.length ||
-          !width ||
-          !height ||
-          buffer.length !== width * height * 4
-        ) {
-          return null;
-        }
-
-        return createImageBitmap(
-          new ImageData(new Uint8ClampedArray(buffer), width, height)
-        );
-      })
-    ).then(([source, highlight]) => {
-      if (active) {
-        setTextureBitmaps({ source, highlight });
-      }
-    });
-
-    return () => {
-      active = false;
-    };
-  }, [rgbaBuffer, uvHighlightRgbaBuffer, width, height]);
-
-  const imgCanvasRef = useRef<HTMLCanvasElement>(null);
-
-  useEffect(() => {
-    let animationFrameId = 0;
-
-    const draw = () => {
-      const canvas = imgCanvasRef.current;
-      if (canvas && textureBitmaps.source) {
-        const context = canvas.getContext('2d');
-        if (context) {
-          context.clearRect(0, 0, width, height);
-          context.globalAlpha = showUvHighlight ? 0.25 : 1;
-          context.filter = `saturate(${showUvHighlight ? '0' : '1'})`;
-          context.drawImage(textureBitmaps.source, 0, 0);
-
-          if (!showUvHighlight) {
-            return;
-          }
-
-          const highlightTextureBitmap =
-            textureBitmaps.highlight ?? textureBitmaps.source;
-          const colorOverlayAlpha = shouldPulseUvColorOverlay
-            ? HOVERED_MODEL_UV_MAX_ALPHA *
-              ((Math.cos(
-                (performance.now() / HOVERED_MODEL_UV_PULSE_MS) * Math.PI * 2
-              ) +
-                1) /
-                2)
-            : HOVERED_MODEL_UV_MAX_ALPHA;
-
-          context.globalAlpha = 1;
-          context.filter = 'saturate(1)';
-
-          context.save();
-
-          context.translate(canvas.width / 2, canvas.height / 2);
-          context.rotate((90 * Math.PI) / 180);
-          context.translate(-canvas.width / 2, -canvas.height / 2);
-
-          uvClipPathGroups.forEach(({ paths, color }) => {
-            paths.forEach((points) => {
-              if (points.length === 0) {
-                return;
-              }
-
-              context.save();
-              context.beginPath();
-              context.moveTo(points[0].x, points[0].y);
-
-              for (let i = 1; i < points.length; i++) {
-                context.lineTo(points[i].x, points[i].y);
-              }
-
-              context.closePath();
-              context.clip();
-
-              context.translate(canvas.width / 2, canvas.height / 2);
-              context.rotate((-90 * Math.PI) / 180);
-              context.translate(-canvas.width / 2, -canvas.height / 2);
-
-              context.drawImage(highlightTextureBitmap, 0, 0);
-
-              if (color) {
-                const [red, green, blue] = color;
-                context.fillStyle = `rgba(${red}, ${green}, ${blue}, ${colorOverlayAlpha})`;
-                context.fillRect(0, 0, canvas.width, canvas.height);
-              }
-
-              context.restore();
-            });
-          });
-          context.restore();
-        }
-      }
-
-      if (shouldPulseUvColorOverlay) {
-        animationFrameId = requestAnimationFrame(draw);
-      }
-    };
-
-    animationFrameId = requestAnimationFrame(draw);
-
-    return () => cancelAnimationFrame(animationFrameId);
-  }, [
-    textureBitmaps,
-    uvClipPathGroups,
-    showUvHighlight,
-    shouldPulseUvColorOverlay
-  ]);
 
   const isSelectable = contentViewMode === 'textures' && !selected;
 
@@ -467,13 +336,14 @@ export default function GuiPanelTexture(props: GuiPanelTextureProps) {
   }
 
   const content = (
-    <ImageBufferCanvas
-      rgbaBuffer={!rgbaBuffer?.length ? rgbaBuffer : undefined}
-      ref={rgbaBuffer?.length ? imgCanvasRef : undefined}
+    <GuiPanelTextureUvPreview
+      imageBufferKey={imageBufferKey}
+      uvHighlightBufferKey={uvHighlightBufferKey}
+      uvClipPathGroups={uvClipPathGroups}
+      showUvHighlight={showUvHighlight}
+      shouldPulseUvColorOverlay={shouldPulseUvColorOverlay}
       width={width}
       height={height}
-      alt={`Texture # ${textureIndex}`}
-      className='img'
     />
   );
 
@@ -491,6 +361,7 @@ export default function GuiPanelTexture(props: GuiPanelTextureProps) {
 
   return (
     <Box
+      ref={textureContainerRef}
       className={clsx(
         `mode-${contentViewMode}`,
         isSelectable && 'selectable',
@@ -507,15 +378,53 @@ export default function GuiPanelTexture(props: GuiPanelTextureProps) {
       )}
       {textureDef ? (
         <>
+          <Tooltip title='Texture options'>
+            <IconButton
+              color='primary'
+              aria-haspopup='true'
+              onClick={onOpenTextureMenu}
+              className='texture-overlay-button'
+              sx={{
+                position: 'absolute',
+                top: 'var(--mui-spacing)',
+                right: 0,
+                visibility: textureMenuOpen ? 'hidden' : 'visible'
+              }}
+            >
+              <MoreVertIcon fontSize='small' />
+            </IconButton>
+          </Tooltip>
+          <Tooltip title='Find models this texture is used in'>
+            <IconButton
+              color='primary'
+              aria-haspopup='true'
+              onClick={onOpenModelNavMenu}
+              className='texture-overlay-button'
+              sx={{
+                position: 'absolute',
+                top: 'calc(var(--mui-spacing) * 6)',
+                right: 0,
+                visibility: modelNavMenuOpen ? 'hidden' : 'visible'
+              }}
+            >
+              <AccountTreeIcon fontSize='small' />
+            </IconButton>
+          </Tooltip>
           <GuiPanelTextureMenu
             textureIndex={textureIndex}
             pixelBufferKeys={textureDef.bufferKeys}
             selectedUvClipPaths={selectedTextureEditUvClipPaths}
             onReplaceImageFile={onSelectNewImageFile}
+            anchorEl={textureMenuAnchorEl}
+            open={textureMenuOpen}
+            onClose={onCloseTextureMenu}
           />
           <GuiPanelTextureModelNavMenu
             textureIndex={textureIndex}
             onModelReferenceHover={setHoveredModelReference}
+            anchorEl={modelNavMenuAnchorEl}
+            open={modelNavMenuOpen}
+            onClose={onCloseModelNavMenu}
           />
         </>
       ) : undefined}
