+++ src/components/panel/textures/GuiPanelTextureMenu.tsx
@@ -1,16 +1,7 @@
-import {
-  KeyboardEvent,
-  MouseEvent,
-  useCallback,
-  useMemo,
-  useRef,
-  useState
-} from 'react';
-import MoreVertIcon from '@mui/icons-material/MoreVert';
-import IconButton from '@mui/material/IconButton';
+import { KeyboardEvent, useCallback, useMemo } from 'react';
 import Menu from '@mui/material/Menu';
 import MenuItem from '@mui/material/MenuItem';
-import { Box, Divider, Tooltip } from '@mui/material';
+import { Divider, Tooltip } from '@mui/material';
 import { TextureImageBufferKeys, UvClipPath } from '@/utils/textures';
 import { TextureColorOptions, useTextureOptions } from '@/modules/model-data';
 
@@ -21,27 +12,19 @@ export default function GuiPanelTextureMenu({
   textureIndex,
   pixelBufferKeys,
   selectedUvClipPaths,
-  onReplaceImageFile
+  onReplaceImageFile,
+  anchorEl,
+  open,
+  onClose
 }: {
   textureIndex: number;
   pixelBufferKeys: TextureImageBufferKeys;
   selectedUvClipPaths: UvClipPath[];
   onReplaceImageFile: (file: File | SharedArrayBuffer) => void;
+  anchorEl: HTMLElement | null;
+  open: boolean;
+  onClose: () => void;
 }) {
-  const menuButtonContainerRef = useRef<HTMLDivElement>(null);
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-  const open = Boolean(anchorEl);
-
-  const handleClick = useCallback((event: MouseEvent<HTMLElement>) => {
-    setAnchorEl(
-      menuButtonContainerRef.current?.parentElement ?? event.currentTarget
-    );
-  }, []);
-
-  const handleClose = useCallback(() => {
-    setAnchorEl(null);
-  }, [setAnchorEl]);
-
   const onMenuItemKeyDown = useCallback((event: KeyboardEvent<HTMLElement>) => {
     if (event.key === 'Tab') {
       event.stopPropagation();
@@ -52,7 +35,7 @@ export default function GuiPanelTextureMenu({
     textureIndex,
     pixelBufferKeys,
     onReplaceImageFile,
-    handleClose,
+    onClose,
     !open
   );
 
@@ -76,47 +59,21 @@ export default function GuiPanelTextureMenu({
   );
 
   return (
-    <Box
-      ref={menuButtonContainerRef}
-      sx={{
-        position: 'absolute',
-        top: 0,
-        right: 0,
-        mt: 1,
-        '& .MuiIconButton-root': {
-          opacity: open ? 1 : 0,
-          transition: (t) => `opacity ${t.transitions.duration.short}ms`
-        },
-        '& .MuiIconButton-root svg': {
-          color: 'var(--mui-palette-primary-contrastText)',
-          filter: 'drop-shadow(3px 5px 2px rgb(0 0 0 / 0.8))'
-        }
-      }}
+    <Menu
+      anchorEl={anchorEl}
+      open={open}
+      onClose={onClose}
+      anchorOrigin={MENU_ANCHOR_ORIGIN}
+      transformOrigin={MENU_TRANSFORM_ORIGIN}
+      sx={{ 'li.MuiMenuItem-root > svg': { mr: 2 } }}
     >
-      <IconButton
-        color='primary'
-        aria-haspopup='true'
-        onClick={handleClick}
-        sx={{ visibility: open ? 'hidden' : 'visible' }}
-      >
-        <MoreVertIcon fontSize='small' />
-      </IconButton>
-      <Menu
-        anchorEl={anchorEl}
-        open={open}
-        onClose={handleClose}
-        anchorOrigin={MENU_ANCHOR_ORIGIN}
-        transformOrigin={MENU_TRANSFORM_ORIGIN}
-        sx={{ 'li.MuiMenuItem-root > svg': { mr: 2 } }}
-      >
-        {options}
-        <Divider />
-        <TextureColorOptions
-          textureIndex={textureIndex}
-          variant='menu'
-          selectedUvClipPaths={selectedUvClipPaths}
-        />
-      </Menu>
-    </Box>
+      {options}
+      <Divider />
+      <TextureColorOptions
+        textureIndex={textureIndex}
+        variant='menu'
+        selectedUvClipPaths={selectedUvClipPaths}
+      />
+    </Menu>
   );
 }
